Extract policy-arns to own file in lib/config
